chore: renaming primitives.ts to components.ts
