+++ packages/react/src/components/AmplifyProvider/AmplifyContext.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { defaultTheme, WebTheme } from '@aws-amplify/ui';
 
-import * as primitives from '../../primitives/primitives';
+import * as primitives from '../../primitives/components';
 
 interface AmplifyContextType {
   components: Partial<typeof primitives>;
